Divide classes into packages
Makes the project much more organised and better adheres to OOP practice

Split classes into packages
